--- policyengine_uk/parameters/gov/dwp/childcare_subsidies/tax_free_childcare/age/standard.yaml
@@ -1,12 +1,0 @@
-description: Age thresholds for tax-free childcare eligibility - standard limit
-metadata:
-  unit: year
-  period: year
-  label: Standard age limit for tax-free childcare eligibility
-  reference:
-    - title: Childcare Act
-      href: https://www.legislation.gov.uk/ukdsi/2015/9780111127063
-    - title: Tax-Free Childcare Guidance
-      href: https://www.gov.uk/tax-free-childcare?step-by-step-nav=d78aeaf6-1747-4d72-9619-f16efb4dd89d
-values:
-  2015-01-01: 12
