--- policyengine_uk/parameters/gov/dwp/childcare_subsidies/tax_free_childcare/contribution/disabled_child.yaml
@@ -1,12 +1,0 @@
-description: Maximum yearly contribution limit for disabled child tax-free childcare
-metadata:
-  unit: currency-GBP
-  period: year
-  label: Disabled child yearly maximum contribution
-  reference:
-    - title: Childcare Payments Act
-      href: https://www.legislation.gov.uk/ukdsi/2015/9780111127063
-    - title: Tax-Free Childcare Guidance
-      href: https://www.gov.uk/tax-free-childcare?step-by-step-nav=d78aeaf6-1747-4d72-9619-f16efb4dd89d
-values:
-  2015-01-01: 4000
