--- policyengine_uk/parameters/gov/dwp/childcare_subsidies/tax_free_childcare/income_thresholds.yaml
@@ -1,23 +1,0 @@
-description: Income thresholds for tax-free childcare eligibility
-metadata:
-  period: year
-  label: Income thresholds for tax-free childcare eligibility
-  reference:
-    - title: Childcare Payments Act
-      href: https://www.legislation.gov.uk/ukdsi/2015/9780111127063
-    - title: Tax-Free Childcare Guidance
-      href: https://www.gov.uk/tax-free-childcare?step-by-step-nav=d78aeaf6-1747-4d72-9619-f16efb4dd89d
-
-brackets:
-  - threshold:
-      2015-01-01: 0
-    amount:
-      2015-01-01: 1_331
-  - threshold:
-      2015-01-01: 18
-    amount:
-      2015-01-01: 1_788
-  - threshold:
-      2015-01-01: 21
-    amount:
-      2015-01-01: 2_379
