+++ policyengine_uk/parameters/gov/hmrc/tax_free_childcare/age/disability.yaml
@@ -1,0 +1,12 @@
+description: The HMRC department extends the tax-free childcare program eligibility to children with disabilities up to this age threshold.
+metadata:
+  unit: year
+  period: year
+  label: Tax-free childcare disability age limit
+  reference:
+    - title: Childcare Payments Act 2014 - c. 28 - section 14
+      href: https://www.legislation.gov.uk/ukpga/2014/28/notes/division/6/4/1
+    - title: Tax-free childcare guidance
+      href: https://www.gov.uk/tax-free-childcare?step-by-step-nav=d78aeaf6-1747-4d72-9619-f16efb4dd89d
+values:
+  2015-01-01: 17
