+++ policyengine_uk/tests/policy/baseline/gov/hmrc/tax_free_childcare/tax_free_childcare_eligibility.yaml
@@ -7,7 +7,7 @@
     tax_free_childcare_incompatibilities_childcare_eligible: true
     tax_free_childcare_work_condition: true
   output:
-    tax_free_childcare_overall_eligible: true
+    tax_free_childcare_eligible: true
 
 - name: Fails age condition only
   period: 2025
@@ -17,7 +17,7 @@
     tax_free_childcare_incompatibilities_childcare_eligible: true
     tax_free_childcare_work_condition: true
   output:
-    tax_free_childcare_overall_eligible: false
+    tax_free_childcare_eligible: false
 
 - name: Fails income condition only
   period: 2025
@@ -27,7 +27,7 @@
     tax_free_childcare_incompatibilities_childcare_eligible: true
     tax_free_childcare_work_condition: true
   output:
-    tax_free_childcare_overall_eligible: false
+    tax_free_childcare_eligible: false
 
 - name: Fails work condition only
   period: 2025
@@ -37,7 +37,7 @@
     tax_free_childcare_incompatibilities_childcare_eligible: true
     tax_free_childcare_work_condition: false
   output:
-    tax_free_childcare_overall_eligible: false
+    tax_free_childcare_eligible: false
 
 - name: Fails all conditions
   period: 2025
@@ -47,41 +47,5 @@
     tax_free_childcare_incompatibilities_childcare_eligible: false
     tax_free_childcare_work_condition: false
   output:
-    tax_free_childcare_overall_eligible: false
+    tax_free_childcare_eligible: false
 
-# Tests for tax_free_childcare
-- name: Eligible standard child
-  period: 2025
-  input:
-    tax_free_childcare_overall_eligible: true
-    is_child: true
-    is_disabled_for_benefits: false
-  output:
-    tax_free_childcare: 2000
-
-- name: Eligible disabled child
-  period: 2025
-  input:
-    tax_free_childcare_overall_eligible: true
-    is_child: true
-    is_disabled_for_benefits: true
-  output:
-    tax_free_childcare: 4000
-
-- name: Not eligible due to not being child
-  period: 2025
-  input:
-    tax_free_childcare_overall_eligible: true
-    is_child: false
-    is_disabled_for_benefits: false
-  output:
-    tax_free_childcare: 0
-
-- name: Not eligible despite being disabled child
-  period: 2025
-  input:
-    tax_free_childcare_overall_eligible: false
-    is_child: true
-    is_disabled_for_benefits: true
-  output:
-    tax_free_childcare: 0
