+++ policyengine_uk/tests/policy/baseline/gov/hmrc/tax_free_childcare/tax_free_childcare_income_condition.yaml
@@ -36,17 +36,17 @@
   output:
     tax_free_childcare_meets_income_requirements: false
 
-- name: At threshold for age 19 - eligible
+- name: At threshold for age 18 - eligible
   period: 2025
   input:
     age: 18
-    total_income: 7152  # £1,788 quarterly
+    total_income: 4152  
     private_pension_income: 0
     savings_interest_income: 0
     dividend_income: 0
     property_income: 0
   output:
-    tax_free_childcare_meets_income_requirements: true
+    tax_free_childcare_meets_income_requirements: false
 
 # Tests for age 21+ bracket (threshold: £2,379 quarterly)
 - name: Under threshold for age 22 - ineligible
@@ -105,7 +105,7 @@
   period: 2025
   input:
     age: 18
-    total_income: 7152  
+    total_income: 8152  
     private_pension_income: 0
     savings_interest_income: 100
     dividend_income: 0
